+++ NEWS.md
@@ -7,6 +7,7 @@ None
 1. Fix #484 - Add function `util_log_ts()`
 2. Fix #485 - Add function `util_singlediff_ts()`
 3. Fix #486 - Add function `util_doublediff_ts()`
+4. Fix #487 - Add function `util_difflog_ts()`
 
 ## Minor Fixes and Improvements
 1. Fix #480 - Add attributes to output of `ts_growth_rate_vec()`